h_signature: Use double rather than long double arithmetic
This speeds up the h_signature by a factor of roughly 2 on an 8th
generation Core i7 CPU. Profiling shows that h_signature calculation
is the single most expensive thing in the code, so this helps a lot.

HOWEVER, some work may need to be done to diagonalize the calculation of
the numerator and denominator of Al to avoid numerical extremes. It
appears that the use of long double was intended to provide some
robustness to that.

+++ include/teb_local_planner/equivalence_relations.h
@@ -63,7 +63,7 @@ namespace teb_local_planner
 class EquivalenceClass
 {
 public:
-   typedef std::complex<long double> cplx;
+   typedef std::complex<double> cplx;
 
    /**
     * @brief Default constructor
